requirements.txt
streamlit
chromadb
sentence-transformers
torch
torchvision
openai
openai-whisper
PyPDF2
python-docx
Pillow
python-dotenv
tqdm
streamlit-webrtc
av
numpy
soundfile